added basic map to game
I tested with the tileset.png from holloween game, created the map.json
using Tiled tool. In the preload() in main.js, load the tileset.png and
the map.json. After that, create the tilemap and link the tileset image
to it then you can create a 'layer' for rendering