added login functionality
Restructured public/:
I moved the fonts and the purple background to the
res/ and the css/ folder, these two folders are for those might be
needed in both pages.
Added login:
now when the player connected to /muzikuro, the #login div is shown on
the top, and the game is loaded in the background. The "requestPlayer"
is emitted when the player entered a name and clicked "Join Game".