restructured project
Scenes are all moved into /client/scenes/*.js
Now the player informations are hold by the Game object and can be
access from the scenes using this.game.players. The socket object is
also moved to a property of the Game Object.
A constants.js has been created to place all the constants (numbers,
strings)